move funcs to separate module
