+++ scivision/catalog/__init__.py
@@ -1,1 +1,2 @@
-from .catalog import keys, query, values, add_dataset, add_model
+from .catalog import keys, query, values
+from .update import add_dataset, add_model
